Moving test to the correct class
